Reestruturacao do projeto para mais organizacao, finalizacao do read_parse
