+++ src/read_parse.c
@@ -1,0 +1,232 @@
+// Include
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <ctype.h>
+#include <string.h>
+#include <readline/readline.h>           //Biblioteca para suportar comandos de um terminal real, como seta pra cima ou pra baixo, historico de comandos, ctrl a, ctrl e, etc
+#include <readline/history.h>
+
+#include "read_parse.h"
+
+// Functions to read and parse
+
+CommandLine* reading(){
+
+    char *linha = readline(">>> ");
+
+    if (linha == NULL) {
+        // Ctrl+D
+        printf("\n");
+        exit(0);
+    }
+    if (strlen(linha) > 0) {
+        add_history(linha);  // adiciona ao historico se nao for vazia
+    }
+
+    CommandLine *cmd_data = parse_line(linha);
+
+    free (linha);
+
+    return cmd_data;
+}
+
+CommandLine* parse_line(const char *line) {
+
+    if (line == NULL || strlen(line) == 0) {
+        return NULL;
+    }
+    // 1. Alocar a estrutura CommandLine
+    CommandLine *cmd_line = malloc(sizeof(CommandLine));
+    if (!cmd_line) {
+        perror("Falha ao alocar CommandLine");
+        return NULL;
+    }
+    cmd_line->commands = NULL;
+    cmd_line->num_commands = 0;
+
+    int num_pipes = 0;
+    for (const char *p = line; (p = strchr(p, '|')) != NULL; p++) {  // contagem de pipes, p recebe um ponteiro para o primeiro |, strchar retorna null quando nao existem mais |
+        num_pipes++;
+    }
+    int num_cmds = num_pipes + 1;
+
+    if (num_cmds == 0 && strlen(line) > 0) {                   // sem pipe mas ainda existe um comando
+        num_cmds = 1;
+    } else if (num_cmds == 0 && strlen(line) == 0) {           //comando vazio
+        free(cmd_line);
+
+        return NULL;
+    }
+
+    cmd_line->num_commands = num_cmds;
+    cmd_line->commands = malloc(num_cmds * sizeof(ParsedCommand));
+
+    if (!cmd_line->commands) {
+        perror("Falha ao alocar array de ParsedCommand");
+        free(cmd_line);
+        return NULL;
+    }
+
+    char *line_copy = strdup(line); // Trabalhar em uma cópia para strtok_r ou strtok
+
+    if (!line_copy) {
+        perror("strdup_line_copy");
+        free(cmd_line->commands);
+        free(cmd_line);
+        return NULL;
+    }
+
+    char *saveptr; // Para strtok_r, se preferir (mais seguro)
+    int i = 0;
+
+    char *segment = strtok_r(line_copy, "|", &saveptr);
+
+    while (segment != NULL && i < cmd_line->num_commands) {
+
+        char *current_segment = segment;
+        while (isspace((unsigned char)*current_segment)) {
+            current_segment++;
+        }
+        char *end = current_segment + strlen(current_segment) - 1;
+        while (end > current_segment && isspace((unsigned char)*end)) {
+            end--;
+        }
+        *(end + 1) = '\0'; // Termina a string após o último caractere não-espaço
+
+        if (strlen(current_segment) == 0) {                  // Se o usuario digitar cmd1 | | cmd 2
+            segment = strtok_r(NULL, "|", &saveptr); // Pega o próximo
+            continue;
+        }
+
+        cmd_line->commands[i].args = parsing_commands(current_segment);
+
+        if (cmd_line->commands[i].args == NULL) {
+            // Erro na tokenização do segmento, liberar tudo
+            for (int k = 0; k < i; k++) {
+                for(int j=0; cmd_line->commands[k].args[j] != NULL; j++) free(cmd_line->commands[k].args[j]);
+                free(cmd_line->commands[k].args);
+            }
+            free(cmd_line->commands);
+            free(cmd_line);
+            free(line_copy);
+            return NULL;
+        }
+        i++;
+        segment = strtok_r(NULL, "|", &saveptr);
+    }
+    cmd_line->num_commands = i;
+
+    free(line_copy);
+    return cmd_line;
+}
+
+char **parsing_commands (char *input){
+
+    char **args = malloc(50 * sizeof(char *));       // Lista de argumentos
+
+    if (!args) {
+        perror("Falha ao alocar argumentos!");
+        return NULL;
+    }
+    const char *delimiters = " \t\n";
+    char *token = strtok(input, delimiters);        // " \t\n" // strtok encontra o delimitador e substitui ele por \0
+
+    int i = 0;
+
+    while (token != NULL && i < 49) {
+
+        args[i] = strdup(token);                  //Copia a string até o primeiro \0 para args
+
+        if (!args[i]) {
+            perror("Falha ao duplicar token!");
+            for (int k = 0; k < i; k++) {
+                free(args[k]);
+            }
+            free(args);
+            return NULL;
+        }
+
+        token = strtok(NULL, delimiters);
+        i++;
+    }
+
+    args[i] = NULL;
+
+    //  for (int i = 0; args[i] != NULL; i++){          // para testes
+    //     printf("string [%d] = %s",i, args[i]);
+    //  }
+    //  fflush(stdout);
+
+    return args;
+
+}
+
+
+// DEBUG
+
+void print_command_line_details(const CommandLine *cmd_line) {
+    if (cmd_line == NULL) {
+        printf("DEBUG: CommandLine é NULL.\n");
+        fflush(stdout); // Garante que a saída seja impressa imediatamente
+        return;
+    }
+
+    printf("DEBUG: --- Detalhes da CommandLine ---\n");
+    printf("DEBUG: Número total de comandos no pipeline: %d\n", cmd_line->num_commands);
+
+    if (cmd_line->num_commands == 0) {
+        printf("DEBUG: Nenhum comando para exibir (num_commands é 0).\n");
+        printf("DEBUG: --- Fim dos Detalhes da CommandLine ---\n");
+        fflush(stdout);
+        return;
+    }
+
+    if (cmd_line->commands == NULL) {
+        printf("DEBUG: Array 'commands' é NULL, embora num_commands seja %d (estado inconsistente!).\n", cmd_line->num_commands);
+        printf("DEBUG: --- Fim dos Detalhes da CommandLine ---\n");
+        fflush(stdout);
+        return;
+    }
+
+    for (int i = 0; i < cmd_line->num_commands; i++) {
+        printf("DEBUG:   Comando %d:\n", i); // Usando índice 0-based para consistência com arrays
+
+        if (cmd_line->commands[i].args == NULL) {
+            printf("DEBUG:     args para o Comando %d é NULL.\n", i);
+            continue; // Passa para o próximo comando no pipeline
+        }
+
+        char **current_args = cmd_line->commands[i].args;
+        for (int j = 0; ; j++) { // Loop infinito, quebra quando encontrar o NULL
+            if (current_args[j] != NULL) {
+                printf("DEBUG:     args[%d]: \"%s\"\n", j, current_args[j]);
+            } else {
+                printf("DEBUG:     args[%d]: (NULL)\n", j); // Imprime a representação do NULL terminador
+                break; // Sai do loop de argumentos para este comando
+            }
+        }
+    }
+    printf("DEBUG: --- Fim dos Detalhes da CommandLine ---\n");
+    fflush(stdout); // Garante que a saída seja impressa imediatamente
+}
+
+// free_memory
+
+void free_command_line(CommandLine *cmd_line) {
+
+    if (!cmd_line) {
+        return;
+    }
+    for (int i = 0; i < cmd_line->num_commands; i++) {
+        if (cmd_line->commands[i].args) {
+            char **arg_ptr = cmd_line->commands[i].args;
+            for (int j = 0; arg_ptr[j] != NULL; j++) {
+                free(arg_ptr[j]);
+            }
+            free(cmd_line->commands[i].args);
+        }
+    }
+    free(cmd_line->commands);
+    free(cmd_line);
+}
