+++ src/read_parse.h
@@ -1,0 +1,25 @@
+#ifndef READ_PARSE_H           // BIBLIOTECA PARA I/O
+#define READ_PARSE_H
+
+// Struct
+
+typedef struct {
+    char **args; // Argumentos para execvp (ex: {"ls", "-l", NULL})
+} ParsedCommand;
+
+typedef struct {
+    ParsedCommand *commands; // Array de comandos
+    int num_commands;        // Número de comandos no pipeline (1 se não for pipeline)
+} CommandLine;
+
+// Prototype
+
+CommandLine* reading ();
+CommandLine* parse_line(const char *line);
+char **parsing_commands (char *input);
+void print_command_line_details(const CommandLine *cmd_line);
+void free_command_line(CommandLine *cmd_line);
+
+
+
+#endif
